dependency inject razorengine and update tests

+++ src/Domain.LinnApps/PurchaseOrders/PurchaseOrderDetail.cs
@@ -50,7 +50,7 @@
 
         public int? OriginalOrderNumber { get; set; }
 
-        public int? OurQty { get; set; }
+        public decimal? OurQty { get; set; }
 
         public decimal? OrderQty { get; set; }
